requirements.txt

termcolor
pyfiglet
python-docx 
colorama

